added clientInfo command
